Cleanup test code before splitting.

+++ mono/tests/tailcall-interface.cs
@@ -20,10 +20,10 @@ interface I1
 	void perturb_interface_offset1 ( );
 
 	[MethodImpl (NoInlining)]
-	int F1 (I2 i2, int counter, long initial_stack, long current_stack = 0);
+	int F1 (I2 i2, long counter, long initial_stack, long current_stack = 0);
 
 	[MethodImpl (NoInlining)]
-	int GF1<TF> (I2 i2, int counter, long initial_stack, long current_stack = 0);
+	int GF1<TF> (I2 i2, long counter, long initial_stack, long current_stack = 0);
 }
 
 interface GI1<TC>
@@ -32,13 +32,13 @@ interface GI1<TC>
 	void perturb_interface_offset2 ( );
 
 	[MethodImpl (NoInlining)]
-	int F1 (GI2<TC> i2, int counter, long initial_stack, long current_stack = 0);
+	int F1 (GI2<TC> i2, long counter, long initial_stack, long current_stack = 0);
 
 	[MethodImpl (NoInlining)]
-	int GF1<TF> (GI2<TC> i2, int counter, long initial_stack, long current_stack = 0);
+	int GF1<TF> (GI2<TC> i2, long counter, long initial_stack, long current_stack = 0);
 
 	[MethodImpl (NoInlining)]
-	int HF1<TF> (GI2<TF> i2, int counter, long initial_stack, long current_stack = 0);
+	int HF1<TF> (GI2<TF> i2, long counter, long initial_stack, long current_stack = 0);
 }
 
 interface I2
@@ -48,10 +48,10 @@ interface I2
 	void perturb_interface_offset3 ( );
 
 	[MethodImpl (NoInlining)]
-	int F2 (I1 i1, int counter, long initial_stack, long current_stack = 0);
+	int F2 (I1 i1, long counter, long initial_stack, long current_stack = 0);
 
 	[MethodImpl (NoInlining)]
-	int GF2<TF> (I1 i1, int counter, long initial_stack, long current_stack = 0);
+	int GF2<TF> (I1 i1, long counter, long initial_stack, long current_stack = 0);
 }
 
 interface GI2<TC>
@@ -62,13 +62,13 @@ interface GI2<TC>
 	void perturb_interface_offset4 ( );
 
 	[MethodImpl (NoInlining)]
-	int F2 (GI1<TC> i1, int counter, long initial_stack, long current_stack = 0);
+	int F2 (GI1<TC> i1, long counter, long initial_stack, long current_stack = 0);
 
 	[MethodImpl (NoInlining)]
-	int GF2<TF> (GI1<TC> i1, int counter, long initial_stack, long current_stack = 0);
+	int GF2<TF> (GI1<TC> i1, long counter, long initial_stack, long current_stack = 0);
 
 	[MethodImpl (NoInlining)]
-	int HF2<TF> (GI1<TF> i1, int counter, long initial_stack, long current_stack = 0);
+	int HF2<TF> (GI1<TF> i1, long counter, long initial_stack, long current_stack = 0);
 }
 
 unsafe public class C1 : I1
@@ -93,7 +93,7 @@ unsafe public class C1 : I1
 	}
 
 	[MethodImpl (NoInlining)]
-	int I1.F1 (I2 i2, int counter, long initial_stack, long current_stack)
+	int I1.F1 (I2 i2, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -102,7 +102,7 @@ unsafe public class C1 : I1
 	}
 
 	[MethodImpl (NoInlining)]
-	int I1.GF1<TF> (I2 i2, int counter, long initial_stack, long current_stack)
+	int I1.GF1<TF> (I2 i2, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -122,7 +122,7 @@ unsafe public class GC1<TC> : GI1<TC>
 	}
 
 	[MethodImpl (NoInlining)]
-	int GI1<TC>.F1 (GI2<TC> i2, int counter, long initial_stack, long current_stack)
+	int GI1<TC>.F1 (GI2<TC> i2, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -131,7 +131,7 @@ unsafe public class GC1<TC> : GI1<TC>
 	}
 
 	[MethodImpl (NoInlining)]
-	int GI1<TC>.GF1<TF> (GI2<TC> i2, int counter, long initial_stack, long current_stack)
+	int GI1<TC>.GF1<TF> (GI2<TC> i2, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -140,7 +140,7 @@ unsafe public class GC1<TC> : GI1<TC>
 	}
 
 	[MethodImpl (NoInlining)]
-	int GI1<TC>.HF1<TF> (GI2<TF> i2, int counter, long initial_stack, long current_stack)
+	int GI1<TC>.HF1<TF> (GI2<TF> i2, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -161,7 +161,7 @@ unsafe public class C2 : I2
 	}
 
 	[MethodImpl (NoInlining)]
-	int I2.F2 (I1 i1, int counter, long initial_stack, long current_stack)
+	int I2.F2 (I1 i1, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -170,7 +170,7 @@ unsafe public class C2 : I2
 	}
 
 	[MethodImpl (NoInlining)]
-	int I2.GF2<TF> (I1 i1, int counter, long initial_stack, long current_stack)
+	int I2.GF2<TF> (I1 i1, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -192,7 +192,7 @@ unsafe public class GC2<TC> : GI2<TC>
 	}
 
 	[MethodImpl (NoInlining)]
-	int GI2<TC>.F2 (GI1<TC> i1, int counter, long initial_stack, long current_stack)
+	int GI2<TC>.F2 (GI1<TC> i1, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -201,7 +201,7 @@ unsafe public class GC2<TC> : GI2<TC>
 	}
 
 	[MethodImpl (NoInlining)]
-	int GI2<TC>.GF2<TF> (GI1<TC> i1, int counter, long initial_stack, long current_stack)
+	int GI2<TC>.GF2<TF> (GI1<TC> i1, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -210,7 +210,7 @@ unsafe public class GC2<TC> : GI2<TC>
 	}
 
 	[MethodImpl (NoInlining)]
-	int GI2<TC>.HF2<TF> (GI1<TF> i1, int counter, long initial_stack, long current_stack)
+	int GI2<TC>.HF2<TF> (GI1<TF> i1, long counter, long initial_stack, long current_stack)
 	{
 		int local;
 		if (counter > 0)
@@ -226,19 +226,19 @@ public class B { }
 interface IC
 {
 	[MethodImpl (NoInlining)]
-	T cast1<T> (object o, int counter = 100, long stack = 0);
+	T cast1<T> (object o, long counter = 100, long stack = 0);
 
 	[MethodImpl (NoInlining)]
-	B cast2 (object o, int counter = 100, long stack = 0);
+	B cast2 (object o, long counter = 100, long stack = 0);
 
 	[MethodImpl (NoInlining)]
-	T cast3<T> (object o, int counter = 100, long stack = 0);
+	T cast3<T> (object o, long counter = 100, long stack = 0);
 
 	[MethodImpl (NoInlining)]
-	B[] cast4 (object o, int counter = 100, long stack = 0);
+	B[] cast4 (object o, long counter = 100, long stack = 0);
 
 	[MethodImpl (NoInlining)]
-	T[] cast5<T> (object o, int counter = 100, long stack = 0);
+	T[] cast5<T> (object o, long counter = 100, long stack = 0);
 }
 
 unsafe public class C
@@ -250,7 +250,7 @@ unsafe public class C
 	}
 
 	[MethodImpl (NoInlining)]
-	public T cast1<T> (object o, int counter = 100, long stack = 0)
+	public T cast1<T> (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -260,7 +260,7 @@ unsafe public class C
 	}
 
 	[MethodImpl (NoInlining)]
-	public B cast2 (object o, int counter = 100, long stack = 0)
+	public B cast2 (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -270,7 +270,7 @@ unsafe public class C
 	}
 
 	[MethodImpl (NoInlining)]
-	public T cast3<T> (object o, int counter = 100, long stack = 0)
+	public T cast3<T> (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -280,7 +280,7 @@ unsafe public class C
 	}
 
 	[MethodImpl (NoInlining)]
-	public B[] cast4 (object o, int counter = 100, long stack = 0)
+	public B[] cast4 (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -290,7 +290,7 @@ unsafe public class C
 	}
 
 	[MethodImpl (NoInlining)]
-	public T[] cast5<T> (object o, int counter = 100, long stack = 0)
+	public T[] cast5<T> (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -309,7 +309,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public static T cast1<T> (object o, int counter = 100, long stack = 0)
+	public static T cast1<T> (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -319,7 +319,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public B cast2 (object o, int counter = 100, long stack = 0)
+	public B cast2 (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -329,7 +329,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public T cast3<T> (object o, int counter = 100, long stack = 0)
+	public T cast3<T> (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -339,7 +339,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public B[] cast4 (object o, int counter = 100, long stack = 0)
+	public B[] cast4 (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -349,7 +349,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public T[] cast5<T> (object o, int counter = 100, long stack = 0)
+	public T[] cast5<T> (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -359,7 +359,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public T1 cast6 (object o, int counter = 100, long stack = 0)
+	public T1 cast6 (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -369,7 +369,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public T1 cast7<T> (object o, int counter = 100, long stack = 0)
+	public T1 cast7<T> (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -379,7 +379,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public T1[] cast8 (object o, int counter = 100, long stack = 0)
+	public T1[] cast8 (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
@@ -389,7 +389,7 @@ unsafe public class D<T1>
 	}
 
 	[MethodImpl (NoInlining)]
-	public T1[] cast9<T> (object o, int counter = 100, long stack = 0)
+	public T1[] cast9<T> (object o, long counter = 100, long stack = 0)
 	{
 		int local;
 		if (counter > 0)
